requirements/base.txt
Fraction~=1.5.1
Pillow~=8.3.1  # https://github.com/python-pillow/Pillow
argon2-cffi~=20.1.0  # https://github.com/hynek/argon2_cffi
channels~=3.0.4
cifrazia-django-jet~=1.1.4
coveralls~=3.2.0
django-allauth~=0.45.0  # https://github.com/pennersr/django-allauth
django-anymail[mailgun]~=8.4  # https://github.com/anymail/django-anymail
django-compressor~=2.4.1  # https://github.com/django-compressor/django-compressor
django-cors-headers~=3.7.0 # https://github.com/adamchainz/django-cors-headers
django-crispy-forms~=1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-environ~=0.4.5  # https://github.com/joke2k/django-environ
django-filter~=2.4.0
django-mjml[requests]~=0.11.0
django-model-utils~=4.1.1  # https://github.com/jazzband/django-model-utils
django-phonenumber-field~=5.2.0
django-storages[google]~=1.11.1  # https://github.com/jschneier/django-storages
djangorestframework-datatables~=0.6.0
djangorestframework~=3.12.4  # https://github.com/encode/django-rest-framework
django~=3.2.6  # pyup: < 3.2  # https://www.djangoproject.com/
flower~=0.9.7  # https://github.com/mher/flower
google-api-core~=1.31.1
google-api-python-client~=2.15.0
google-auth~=1.34.0
google-cloud-secret-manager~=2.7.0
google-cloud-storage~=1.41.1
openpyxl~=3.0.7
phonenumbers~=8.12.29
python-slugify~=5.0.2  # https://github.com/un33k/python-slugify
pytz~=2021.1  # https://github.com/stub42/pytz
rcssmin~=1.0.6  # https://github.com/ndparker/rcssmin
requests~=2.26.0
uvicorn[standard]~=0.14.0  # https://github.com/encode/uvicorn
whitenoise~=5.3.0  # https://github.com/evansd/whitenoise
